feat: Extract WrappedTextView and add collapse test  - Enhance WrappedTextView to support maxVisibleLines and expanded: add fields (int maxVisibleLines = 2, boolean expanded = false) and setters. Update measure() and paintComponent() so when not expanded and lines.size() > maxVisibleLines, only the first maxVisibleLines are considered for contentHeight, and the last visible line is rendered with an ellipsis appended. Implement a helper to trim the last visible line so that " …" fits within availableWidth (using FontMetrics). [Verify: new unit test WrappedTextViewCollapseTest renders two lines and ellipsis for long text; contentHeight equals 2*lineHeight when collapsed].
